Fix views, add recyclerView for Results
Change the views to RelativeLayout, add the clear fab button.
Added recyclerViews to display the 15 results if needed, and also making them clickcable.
Created the Utils folder for handling Landmarks formatting, instead of coordinates, it gives city and country of the landmark.
Also it gives a right percentage value on the confidence.
Added loading for retrieving results.

TODO STILL: Handling uploading image and camera photo on save instance state