Add icons and Rename and Refactor
Added add and check icons. Tried to implement animation. when click on fab button to add imageInfo to database the plus sign
supposedly turn into a check sign. There's an issue where when user scrolls up after reading the text, the fab button shows but when
tapped again, the second button state doesn't show up. It's a Google bug that's only on the 28 version.

Renamed CameraUtils to PermissionUtils where it deals with camera and gallery permissions.

Refactor some of the code.