Fix Fab button issue
I have a collapsing toolbar layout and a FAB. I dynamically change the FAB icon on click. It works wells. But when the toolbar collapses and expanded again. Clicking on the FAB makes the icon disappear. When I collapse and expanded again, the icon appears, and the icon is changed!
SOLUTION: implement a changeFab function where it receives a resource it hides set the resouce to fab button and set to show it again