Finish implementing widget
User can add widget and see the last three landmarks that they added to the library.
For now they can click it will take to the MyLandmarks Activity.
Next iteration: check if user is logged in, and once user click in the individual widget,
they will be able to see the landmark details